requirements.txt
flair
torch
pandas
numpy
keras
glove_python